ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.18.1
paddlepaddle==2.6.0
paddleocr==2.7.3
pillow==10.2.0
pdf2image==1.17.0
pymupdf==1.23.26
boto3==1.34.20
httpx==0.26.0
structlog==24.1.0
python-multipart==0.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
